+++ base/deprecated.jl
@@ -1485,6 +1485,17 @@ end
 using .DSP
 export conv, conv2, deconv, filt, filt!, xcorr
 
+function replace(s::AbstractString, pat, f, n::Integer)
+    if n <= 0
+        depwarn(string("`replace(s, pat, r, count)` with `count <= 0` is deprecated, use ",
+                       "`replace(s, pat, r, typemax(Int))` or replace(s, pat, r)` instead"),
+                :replace)
+        Base.replace(s, pat, f)
+    else
+        Base.replace_new(String(s), pat, f, Int(n))
+    end
+end
+
 # END 0.7 deprecations
 
 # BEGIN 1.0 deprecations